catchsensor

+++ Project_wsl_2026/Software/system_core.ver1/main_tennsy4.0/lib/Kicker/kicker.cpp
@@ -55,7 +55,7 @@ void catchsensor_init(int pin)
     pinMode(catchsensor_pin, INPUT);
 }
 
-unsigned int get_catchsensor()
+uint8_t get_catchsensor()
 {
     return digitalRead(catchsensor_pin);
 }
